+++ changelog.md
@@ -1,6 +1,7 @@
 ### 0.0.6 - [Sep 2026]
 - **[change]** - improve input validation
 - **[change]** - simplify fill-down and validate inputs
+- **[change]** - check for cycles in JSON input
 - **[fix]** - linting 
 - **[update]** - dependencies